Added AI-less, boundingbox-less, birds.
Same bird, different scales.
Odd graphical bug in one wing, on both models. However the bug is
different depending on scale. Looking at the original .blend a cause cannot be found. The blend file has been uploaded to the meta repo.
Due to lack of documentation the .prefabs and .mat are barren. The bird skin is basic and in need of artistic opinion.